+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 
 ## [Unreleased]
 
+### Added
+- `Forge` subclasses can now grab config via envvars prefixed `FORGE_`. Observe
+  that `Forge` now derives from `pydantic.BaseSettings`, see [BaseSettings
+  docs](https://docs.pydantic.dev/usage/settings/).
+
 
 ## v0.9.0 - 2023-03-12